+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Support examples/control_gallery.rb
 - Support `open_file` and `save_file`
+- Support `quit_menu_item` with `on_clicked` listener
 - Fix issue with always setting menu item text to 'Version' (correctly set to passed argument instead)
 
 ## 0.0.4